requirements.txt
aiogram==3.21.0
aiohttp==3.9.5
pydantic==2.11.7
python-dotenv==1.1.1
